+++ frontend/smarty-pants/services/authService.ts
@@ -158,3 +158,7 @@ class AuthService {
 }
 
 export default AuthService.getInstance();
+
+export interface IAuthService {
+    getAuthenticatedAxiosInstance(): AxiosInstance;
+}
